+++ example/lib/main.dart
@@ -5,6 +5,7 @@ import 'dart:async';
 import 'dart:convert';
 
 import 'package:googlecloudspeech/googlecloudspeech.dart';
+import 'package:file_picker/file_picker.dart';
 import 'package:flutter/services.dart';
 
 
@@ -20,37 +21,61 @@ class _MyAppState extends State<MyApp> {
   // ADD API_KEY HERE
   String _apiKey = '';
 
+  String _audioPath;
+  String _response;
+
   @override
   void initState() {
     super.initState();
+
+    _audioPath = '';
+    _response  = '';
   }
 
   Future<void> getListOfOperations() async {
 
     Map<String, dynamic> responseBody = await GoogleSTTOperations.v1_list({}, _apiKey);
 
-    print("----------");
+    print("---------- RESPONSE ----------");
     print(responseBody);
-    print("----------");
+    print("------------------------------");
   }
 
   Future<void> speechToText() async {
 
     try {
-      // This is an example asset file. Replace this with a file on the device
-      var bytes            = await rootBundle.load('assets/example_file.flac');
-      String base64audio   = base64.encode(Uint8List.view(bytes.buffer));
+//      // This is an example asset file. Replace this with a file on the device
+//      var bytes            = await rootBundle.load('assets/example_file.flac');
+//      String base64audio   = base64.encode(Uint8List.view(bytes.buffer));
 
-      Map<String, dynamic> response = await GoogleSTTSpeech.v1_recognize({}, _apiKey, base64audio);
+      // NOTE: Currently config dictionary supports only file types 'flac' or 'wav'
+      Map<String, dynamic> response = await GoogleSTTSpeech.v1_recognize({}, _apiKey, 'file_path', _audioPath);
 
-      print("------");
+      print("------ RESPONSE ------");
       print(response);
-      print("------");
+      print("----------------------");
 
+      setState(() {
+        _response = response.toString();
+      });
     } catch (e) {
       print(e.toString());
     }
+  }
 
+  void _openFileExplorer() async {
+
+    try {
+      String path = await FilePicker.getFilePath(type: FileType.audio,);
+
+      setState(() {
+        _audioPath = path;
+      });
+
+    } on PlatformException catch (e) {
+      print("Unsupported operation" + e.toString());
+    }
+    if (!mounted) return;
   }
 
   @override
@@ -77,10 +102,36 @@ class _MyAppState extends State<MyApp> {
                     highlightColor : Colors.grey[600],
                     shape          : RoundedRectangleBorder(borderRadius: BorderRadius.circular(10.0)),
 
-                    child : Text('Operations List', style : TextStyle(fontSize : 25, fontWeight : FontWeight.bold, color : Colors.grey[500])),
+                    child : Text('Select File', style : TextStyle(fontSize : 25, fontWeight : FontWeight.bold, color : Colors.grey[500])),
+                    onPressed: _openFileExplorer,
+                  ),
+                ),
+                Container(
+                  padding: EdgeInsets.fromLTRB(10, 10, 10, 10),
+
+                  child: Text(_audioPath, style: TextStyle(fontSize: 12))
+                ),
+                Container(
+                  height : 60,
+                  width  : 270,
+                  decoration : BoxDecoration(
+                    borderRadius : BorderRadius.circular(10.0),
+                    color        : Colors.white,
+                  ),
+
+                  child : MaterialButton(
+                    highlightColor : Colors.grey[600],
+                    shape          : RoundedRectangleBorder(borderRadius: BorderRadius.circular(10.0)),
+
+                    child : Text('Speech 2 Text', style : TextStyle(fontSize : 25, fontWeight : FontWeight.bold, color : Colors.grey[500])),
                     onPressed: () { speechToText(); },
                   ),
                 ),
+                Container(
+                  padding: EdgeInsets.fromLTRB(10, 10, 10, 10),
+
+                  child: Text(_response, style : TextStyle(fontSize: 12))
+                )
               ],
             )
         ),