+++ CHANGELOG.md
@@ -13,3 +13,11 @@
 # 0.1.2
 
 * Code prettified
+
+# 0.1.3
+
+* Convert audio file to base64 string
+* Support other audio file formats
+* Removed (and untrack) system generated files
+* Added LICENSE
+